Move protocol transition tests
