+++ spark/src/test/scala/org/apache/spark/sql/delta/DeltaProtocolVersionSuite.scala
@@ -3517,169 +3517,6 @@ trait DeltaProtocolVersionSuiteBase extends QueryTest
     }
   }
 
-  private def validVersions = Seq((1, 1), (1, 2), (1, 3), (1, 4), (2, 5), (1, 7), (3, 7))
-  private def invalidVersions = Seq((2, 2), (2, 3))
-  for ((readerVersion, writerVersion) <- validVersions ++ invalidVersions)
-  test("Legacy features are added when setting legacy versions: " +
-      s"readerVersionToSet = $readerVersion, writerVersionToSet = $writerVersion") {
-    withTempDir { dir =>
-      val deltaLog = DeltaLog.forTable(spark, dir)
-
-      // Creates a table with (3, 7) versions with the given table feature.
-      createTableWithFeature(
-        deltaLog,
-        TestRemovableWriterFeature,
-        TestRemovableWriterFeature.TABLE_PROP_KEY)
-
-      sql(
-        s"""
-           |ALTER TABLE delta.`${deltaLog.dataPath}` SET TBLPROPERTIES (
-           |  'delta.minReaderVersion' = $readerVersion,
-           |  'delta.minWriterVersion' = $writerVersion
-           |)""".stripMargin)
-
-      val expected = Protocol(readerVersion, writerVersion).implicitlySupportedFeatures ++
-        Set(InvariantsTableFeature, AppendOnlyTableFeature, TestRemovableWriterFeature)
-      assert(deltaLog.update().protocol.readerAndWriterFeatureNames === expected.map(_.name))
-    }
-  }
-
-  test("Setting protocol versions in a legacy table should not add any features") {
-    withTempDir { dir =>
-      val deltaLog = DeltaLog.forTable(spark, dir)
-
-      sql(
-        s"""CREATE TABLE delta.`${deltaLog.dataPath}` (id bigint) USING delta
-           |TBLPROPERTIES (
-           |delta.minReaderVersion = '1',
-           |delta.minWriterVersion = '2'
-           |)""".stripMargin)
-
-      sql(
-        s"""
-           |ALTER TABLE delta.`${deltaLog.dataPath}` SET TBLPROPERTIES (
-           |  'delta.minReaderVersion' = '1',
-           |  'delta.minWriterVersion' = '4'
-           |)""".stripMargin)
-
-      // There should be no explicitly added legacy features.
-      assert(deltaLog.update().protocol == Protocol(1, 4))
-    }
-  }
-
-  for {
-    tableFeatureToAdd <- Seq(TestRemovableWriterFeature, TestRemovableReaderWriterFeature)
-    downgradeVersionToSet <- Seq(1, 2, 3, 4, 5, 6)
-    preemptiveVersionDowngrade <- BOOLEAN_DOMAIN
-  } test("Protocol versions are always downgraded to the minimum required " +
-      s"tableFeatureToAdd: ${tableFeatureToAdd.name}, " +
-      s"downgradeVersionToSet: $downgradeVersionToSet, " +
-      s"preemptiveVersionDowngrade: $preemptiveVersionDowngrade") {
-    withTempDir { dir =>
-      val deltaLog = DeltaLog.forTable(spark, dir)
-
-      sql(
-        s"""CREATE TABLE delta.`${deltaLog.dataPath}` (id bigint) USING delta
-           |TBLPROPERTIES (
-           |delta.minReaderVersion = ${Math.max(tableFeatureToAdd.minReaderVersion, 1)},
-           |delta.minWriterVersion = $TABLE_FEATURES_MIN_WRITER_VERSION,
-           |delta.feature.${tableFeatureToAdd.name} = 'supported',
-           |delta.feature.${ChangeDataFeedTableFeature.name} = 'supported'
-           |)""".stripMargin)
-
-      val downgradeProtocolVersionsSQL =
-        s"""
-           |ALTER TABLE delta.`${deltaLog.dataPath}` SET TBLPROPERTIES (
-           |  'delta.minReaderVersion' = 1,
-           |  'delta.minWriterVersion' = $downgradeVersionToSet
-           |)""".stripMargin
-
-      if (preemptiveVersionDowngrade) sql(downgradeProtocolVersionsSQL)
-
-      AlterTableDropFeatureDeltaCommand(
-        DeltaTableV2(spark, deltaLog.dataPath),
-        tableFeatureToAdd.name,
-        truncateHistory = tableFeatureToAdd.isReaderWriterFeature).run(spark)
-
-      if (!preemptiveVersionDowngrade) sql(downgradeProtocolVersionsSQL)
-
-      val expectedProtocol = if (downgradeVersionToSet < 4) {
-        Protocol(tableFeatureToAdd.minReaderVersion, 7).withFeature(ChangeDataFeedTableFeature)
-          .merge(Protocol(1, downgradeVersionToSet))
-      } else {
-        Protocol(1, downgradeVersionToSet)
-      }
-      assert(deltaLog.update().protocol === expectedProtocol)
-    }
-  }
-
-  for {
-    tableFeatureToAdd <- Seq(TestRemovableWriterFeature, TestRemovableReaderWriterFeature)
-    setLegacyVersions <- BOOLEAN_DOMAIN
-    downgradeAfterDrop <- if (setLegacyVersions) BOOLEAN_DOMAIN else Seq(false)
-  } test("SOP for downgrading to legacy protocol versions for tables created with features. " +
-      s"tableFeatureToAdd: ${tableFeatureToAdd.name}, setLegacyVersions: $setLegacyVersions, " +
-      s"downgradeAfterDrop: ${downgradeAfterDrop}") {
-    withTempDir { dir =>
-      val deltaLog = DeltaLog.forTable(spark, dir)
-
-      sql(
-        s"""CREATE TABLE delta.`${deltaLog.dataPath}` (id bigint) USING delta
-           |TBLPROPERTIES (
-           |delta.minReaderVersion = $TABLE_FEATURES_MIN_READER_VERSION,
-           |delta.minWriterVersion = $TABLE_FEATURES_MIN_WRITER_VERSION,
-           |delta.feature.${tableFeatureToAdd.name} = 'supported',
-           |delta.feature.${ChangeDataFeedTableFeature.name} = 'supported'
-           |)""".stripMargin)
-
-      val downgradeProtocolVersionsSQL =
-        s"""
-           |ALTER TABLE delta.`${deltaLog.dataPath}` SET TBLPROPERTIES (
-           |  'delta.minReaderVersion' = 1,
-           |  'delta.minWriterVersion' = 4
-           |)""".stripMargin
-
-
-      if (setLegacyVersions && !downgradeAfterDrop) sql(downgradeProtocolVersionsSQL)
-
-      AlterTableDropFeatureDeltaCommand(
-        DeltaTableV2(spark, deltaLog.dataPath),
-        tableFeatureToAdd.name,
-        truncateHistory = tableFeatureToAdd.isReaderWriterFeature).run(spark)
-
-      if (setLegacyVersions && downgradeAfterDrop) sql(downgradeProtocolVersionsSQL)
-
-      val expectedProtocol = if (setLegacyVersions) {
-        Protocol(1, 4)
-      } else {
-        Protocol(1, TABLE_FEATURES_MIN_WRITER_VERSION).withFeature(ChangeDataFeedTableFeature)
-      }
-      assert(deltaLog.update().protocol === expectedProtocol)
-    }
-  }
-
-  test("ALTER TABLE should downgrade protocol versions when needed") {
-    withTempDir { dir =>
-      val deltaLog = DeltaLog.forTable(spark, dir)
-
-      sql(
-        s"""CREATE TABLE delta.`${deltaLog.dataPath}` (id bigint) USING delta
-           |TBLPROPERTIES (
-           |delta.minReaderVersion = $TABLE_FEATURES_MIN_READER_VERSION,
-           |delta.minWriterVersion = $TABLE_FEATURES_MIN_WRITER_VERSION,
-           |delta.feature.${InvariantsTableFeature.name} = 'supported'
-           |)""".stripMargin)
-
-      sql(
-        s"""
-           |ALTER TABLE delta.`${deltaLog.dataPath}` SET TBLPROPERTIES (
-           |delta.feature.${AppendOnlyTableFeature.name} = 'supported'
-           |)""".stripMargin)
-
-      assert(deltaLog.update().protocol === Protocol(1, 2))
-    }
-  }
-
   private def dropV2CheckpointsTableFeature(spark: SparkSession, log: DeltaLog): Unit = {
     spark.sql(s"ALTER TABLE delta.`${log.dataPath}` DROP FEATURE " +
       s"`${V2CheckpointTableFeature.name}`")
